pyproject.toml
[project]
name = "chimera"
version = "0.1.0"
description = "Project Chimera: Autonomous Influencer Network"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "mcp",
    "weaviate-client>=4.0.0",
    "redis>=5.0.0",
    "requests>=2.32.0",
    "python-dotenv",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "black>=23.0"
]

[tool.setuptools]
packages = ["chimera"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
